(BSR) feat(cheactodes): extract some pages in cheatcodes/pages/others folder
