+++ src/cheatcodes/pages/others/CheatcodesScreenDebugInformations.tsx
