+++ src/cheatcodes/pages/others/CheatcodesScreenNewCaledonia.tsx
