+++ src/features/navigation/CheatcodesStackNavigator/CheatcodesStackNavigator.tsx
@@ -12,15 +12,6 @@ import { CheatcodesScreenHighlightThematicHomeHeader } from 'cheatcodes/pages/fe
 import { CheatcodesNavigationIdentityCheck } from 'cheatcodes/pages/features/identityCheck/CheatcodesNavigationIdentityCheck'
 import { CheatcodesNavigationNewIdentificationFlow } from 'cheatcodes/pages/features/identityCheck/CheatcodesNavigationNewIdentificationFlow'
 import { CheatcodesNavigationInternal } from 'cheatcodes/pages/features/internal/CheatcodesNavigationInternal'
-import { CheatcodesNavigationAccountManagement } from 'cheatcodes/pages/features/other/CheatcodesNavigationAccountManagement'
-import { CheatcodesNavigationErrors } from 'cheatcodes/pages/features/other/CheatcodesNavigationErrors'
-import { CheatcodesNavigationNotScreensPages } from 'cheatcodes/pages/features/other/CheatcodesNavigationNotScreensPages'
-import { CheatcodesNavigationSignUp } from 'cheatcodes/pages/features/other/CheatcodesNavigationSignUp'
-import { CheatcodesScreenAccesLibre } from 'cheatcodes/pages/features/other/CheatcodesScreenAccesLibre'
-import { CheatcodesScreenDebugInformations } from 'cheatcodes/pages/features/other/CheatcodesScreenDebugInformations'
-import { CheatcodesScreenFeatureFlags } from 'cheatcodes/pages/features/other/CheatcodesScreenFeatureFlags'
-import { CheatcodesScreenNewCaledonia } from 'cheatcodes/pages/features/other/CheatcodesScreenNewCaledonia'
-import { CheatcodesScreenRemoteConfig } from 'cheatcodes/pages/features/other/CheatcodesScreenRemoteConfig'
 import { CheatcodesNavigationProfile } from 'cheatcodes/pages/features/profile/CheatcodesNavigationProfile'
 import { CheatcodesNavigationShare } from 'cheatcodes/pages/features/share/CheatcodesNavigationShare'
 import { CheatcodesNavigationSubscription } from 'cheatcodes/pages/features/subscription/CheatcodesNavigationSubscription'
@@ -29,6 +20,15 @@ import { CheatcodesScreenTrustedDeviceInfos } from 'cheatcodes/pages/features/tr
 import { CheatcodesNavigationOnboarding } from 'cheatcodes/pages/features/tutorial/CheatcodesNavigationOnboarding'
 import { CheatcodesNavigationProfileTutorial } from 'cheatcodes/pages/features/tutorial/CheatcodesNavigationProfileTutorial'
 import { CheatcodesNavigationTutorial } from 'cheatcodes/pages/features/tutorial/CheatcodesNavigationTutorial'
+import { CheatcodesNavigationAccountManagement } from 'cheatcodes/pages/others/CheatcodesNavigationAccountManagement'
+import { CheatcodesNavigationErrors } from 'cheatcodes/pages/others/CheatcodesNavigationErrors'
+import { CheatcodesNavigationNotScreensPages } from 'cheatcodes/pages/others/CheatcodesNavigationNotScreensPages'
+import { CheatcodesNavigationSignUp } from 'cheatcodes/pages/others/CheatcodesNavigationSignUp'
+import { CheatcodesScreenAccesLibre } from 'cheatcodes/pages/others/CheatcodesScreenAccesLibre'
+import { CheatcodesScreenDebugInformations } from 'cheatcodes/pages/others/CheatcodesScreenDebugInformations'
+import { CheatcodesScreenFeatureFlags } from 'cheatcodes/pages/others/CheatcodesScreenFeatureFlags'
+import { CheatcodesScreenNewCaledonia } from 'cheatcodes/pages/others/CheatcodesScreenNewCaledonia'
+import { CheatcodesScreenRemoteConfig } from 'cheatcodes/pages/others/CheatcodesScreenRemoteConfig'
 import { withAsyncErrorBoundary } from 'features/errors/hocs/withAsyncErrorBoundary'
 import {
   CheatcodesStackParamList,
